% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_dict.R
\name{create_color_dict}
\alias{create_color_dict}
\title{create color dictionary function}
\usage{
create_color_dict(data, metadata)
}
\arguments{
\item{data}{your otu data without added metadata}

\item{metadata}{your metadata with sample name in the first column and a column named 'site'}
}
\description{
This function generates a color dictionary from a csv file with taxa names in column 1, and six digit numbers for colors in column 2
}
\examples{
create_color_dict(data, metadata)
}
\keyword{color}
