% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{create metadata function}
\usage{
create_metadata(data, date = TRUE)
}
\arguments{
\item{date}{argument to format dates in your metadata, your csv file must have a column named "date", defaults to TRUE}
}
\description{
This function imports your otu table metadata. You must have a csv file with "meta" in the file name in the current working directory.
}
\examples{
metadata <- create_metadata()
}
\keyword{metadata}
