% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_otu_table.R
\name{import_otu_table}
\alias{import_otu_table}
\title{import otu table function}
\usage{
import_otu_table(normalize = TRUE, add_metadata = TRUE,
  format_date = TRUE)
}
\arguments{
\item{normalize}{argument to normalize data to relative abundance, defaults to TRUE}

\item{add_metadata}{argument to normalize data to relative abundance, defaults to TRUE}

\item{format_date}{argument to normalize data to relative abundance, defaults to TRUE}
}
\description{
This function reads your data file, formats sample ID's and taxa names, and converts
absolute abundances to relative abundances. The otu table must contain "otu" in the
file name with sample ID's as column names and taxa as row names.#'the metadata file
must have "meta" in the file name and contain matching sample ID's to the otu table in
the first column. To format date, there must be a column named "date" in the metadata file.
}
\examples{
import_otu_table(normalize=TRUE, add_metadata=TRUE, format_date=TRUE) #returns dataframe of normalized otu table with added metadata
import_otu_table(normalize=FALSE) #returns otu table with absolute abundances
}
\keyword{otu}
\keyword{table}
