% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_communities.R
\name{plot_communities}
\alias{plot_communities}
\title{plot communities function}
\usage{
plot_communities(data, color_dict = NULL, metadata, NMDS = FALSE,
  barplot = FALSE, dendrogram = FALSE, dendro_bar = FALSE,
  interactive = FALSE, threshold_abundance = 5, phylum = FALSE,
  legend = FALSE, perm_val = 100)
}
\description{
This function...
}
\examples{
examples...
}
\keyword{plot}
