% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simper_test.R
\name{simper_test}
\alias{simper_test}
\title{simper test function}
\usage{
simper_test(data, metadata, comparison_index = 2, num_perm = 1,
  p_threshold = 1, sim_plot = TRUE, color_dict = NULL,
  phylum = FALSE)
}
\description{
This function...
}
\examples{
examples...
}
\keyword{simper}
