% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{transform data function}
\usage{
transform_data(data, presence_absence = FALSE, square_root = FALSE,
  one_over_x = FALSE)
}
\arguments{
\item{data}{your otu data without added metadata}

\item{presence_absence}{a method to transform your data to binary, defaults to FALSE}

\item{square_root}{a method to square root transform your data, defaults to FALSE}

\item{one_over_x}{a method to transform your data by dividing 1 by each value, defaults to FALSE}
}
\description{
This function transforms your otu table without metadata using a variety of metrics. only one metric can be true at a time.
}
\examples{
data.pres.abs <- transform_data(data, presence_absence = TRUE)
}
\keyword{transform}
